# gnrc_netif_ieee802154: drop duplicate broadcast packets (optionally)