Can be activated with `USEMODULE += gnrc_netif_dedup` (considering `gnrc_netif` is activated, see #7424). Testing procedures see https://github.com/RIOT-OS/RIOT/pull/7577#issuecomment-448372544

Depends on ~~#7409~~ (merged).