@gebart 2017-09-06
Why only broadcast packets? The dedup is just as useful for unicast, for example when a packet is correctly received, but the ack is not received by the sender.