@miri64 2017-09-06
It makes it a little bit more complicated, but will do.