@zhuoshuguo 2017-09-07
A simple solution maybe: you don't need to count whether the received packet is bcast or not, just record the `source address` and `sequence` information (don't care about destination address), and use them to judge duplication.