@miri64 2017-10-12
Rebased and no longer dependent on any PR, but please DO NOT MERGE yet!