@miri64 2017-10-12
Addressed comments.