@miri64 2017-10-12
Addressed comments. Just a friendly reminder to please not merged, before the whole netif2 situation on the maintainers list was discussed.