@miri64 2017-11-27
Rebased to current master