@gebart 2017-12-08
needs rebase