@miri64 2017-12-08
And adaptation to #7736 ;-)