@gebart 2018-06-11
@miri64 this needs rebase after netif2 was renamed netif. #8198 has been merged