@miri64 2018-06-11
@gebart unless this is a pressing matter to you, I rather would like to wait for the layered netdev adaptation of IEEE 802.15.4 and put this on top of that as an additional layer. @bergzand promised to deliver on that in the very near future.