@miri64 2018-12-18
Rebased and adopted for current master.