@PeterKietzmann 2018-12-18
I did some basic testing:
- *examples/gnrc_networking* on two samr21-xpro boards
- compiled with `USEMODULE=gnrc_netif_dedup`
- enabled debug messages in *gnrc_netif_ieee802154.c*
- disable `autoack` feature on board A
- `ping6` from board A to board B
- look at the following outputs

**current master**
```
# ping6 1 fe80::7b62:1b6d:89cc:89ca
# _recv_ieee802154: received packet from 79:62:1B:6D:89:CC:89:CA of length 15
# 00000000  7A  33  3A  81  00  7C  9B  00  5C  00  01  5C  5C  5C  5C
# _recv_ieee802154: reallocating.
# _recv_ieee802154: received packet from 79:62:1B:6D:89:CC:89:CA of length 15
# 00000000  7A  33  3A  81  00  7C  9B  00  5C  00  01  5C  5C  5C  5C
# _recv_ieee802154: reallocating.
# _recv_ieee802154: received packet from 79:62:1B:6D:89:CC:89:CA of length 15
# 00000000  7A  33  3A  81  00  7C  9B  00  5C  00  01  5C  5C  5C  5C
# _recv_ieee802154: reallocating.
# 12 bytes from fe80::7b62:1b6d:89cc:89ca: id=92 seq=1 hop limit=64 time = 89.876 ms
# dropping additional response packet (probably caused by duplicates)
# dropping additional response packet (probably caused by duplicates)
# --- fe80::7b62:1b6d:89cc:89ca ping statistics ---
# 1 packets transmitted, 1 received, 0% packet loss, time 0.06109588 s
# rtt min/avg/max = 89.876/89.876/89.876 ms

```

**with this PR**
```
#  ping6 1  fe80::7b62:1b6d:89cc:89ca
# _recv_ieee802154: received packet from 79:62:1B:6D:89:CC:89:CA of length 15
# 00000000  7B  33  3A  81  00  8E  B6  00  53  00  01  53  53  53  53
# _recv_ieee802154: reallocating.
# _recv_ieee802154: packet dropped by deduplication
# 12 bytes from fe80::7b62:1b6d:89cc:89ca: id=83 seq=1 hop limit=255 time = 45.776 ms
# --- fe80::7b62:1b6d:89cc:89ca ping statistics ---
# 1 packets transmitted, 1 received, 0% packet loss, time 0.0653424 s
# rtt min/avg/max = 45.776/45.776/45.776 ms
```

-> seems to work! For some reason I don't see all retransmissions I would expect to be sent by board B. Probably some timing issues...