@miri64 2018-12-18
> For some reason I don't see all retransmissions I would expect to be sent by board B. Probably some timing issues...

I'd say this is unrelated to this PR.