@PeterKietzmann 2019-01-30
@miri64, @aabadie, @gebart what's your opinion here? With some form of documentation I'd ACK this PR. Needs rebase btw...