@PeterKietzmann 2019-02-04
I still miss some kind of documentation. How would a user be aware of that module? Don't know what's the best location, maybe doxygen in *net/gnrc/netif/ieee802154.h*?