@miri64 2019-02-04
> I still miss some kind of documentation. How would a user be aware of that module? Don't know what's the best location, maybe doxygen in _net/gnrc/netif/ieee802154.h_?

Now that you mention it, it doesn't really make sense to store the data in `netdev_ieee802154`, since all the implementation logic is happening in `gnrc_netif` (the inclusion of the data even is dependent on **gnrc_**netif). Will fix that as well with the added doc.