@PeterKietzmann 2019-02-11
I dismissed @gebart's review as he's been unavailable for quite some time. @gebart please give me a sign if you disagree with this PR. I'll handle the follow-ups then. Otherwise: ACK and go